move search helpers into filter-fields
