+++ frontend/src/components/filter_fields/helpers.ts
