+++ frontend/src/pages/operation_show/layout/toolbar/index.tsx
@@ -8,7 +8,7 @@ import { Tag, User, ViewName } from 'src/global_types'
 import { default as Button, ButtonGroup } from 'src/components/button'
 import { FilterFieldsGrid } from 'src/components/filter_fields/filter-field-grid'
 import Input from 'src/components/input'
-import { stringToSearch, SearchOptions, stringifySearch } from 'src/components/search_query_builder/helpers'
+import { stringToSearch, SearchOptions, stringifySearch } from 'src/components/filter_fields/helpers'
 import { useWiredData, useModal, renderModals } from 'src/helpers'
 import { getTags, listEvidenceCreators } from 'src/services'
 import { CreateButtonPosition } from '..'
